[IN] Memory leak detector for kernel allocator + smoke test (#11)

+++ hal/source/host/host/source/time.zig
@@ -35,6 +35,6 @@ pub const Time = struct {
     }
 
     pub fn sleep_us(us: u64) void {
-        std.Thread.sleep(u64(us) * 1000);
+        std.Thread.sleep(us * 1000);
     }
 };
